Fix Qiskit dependency in package metadata (#90)
`setup.py` was using its own list dependencies and not the list in
`requirements.txt`. Here `setup.py` is changed to read
`requirements.txt` so that the list is specified only once.

With this change, the intended change from 45eccf601182beea446feb59486d7e600cd77f6c
of changing the Qiskit dependency from `qiskit-terra` to `qiskit` is
also now applied to the package metadata.

Additionally, the dependency on `typing-extensions` was removed because
it was only needed for Python 3.7 and `setup.py` already specified
`python_requires>=3.8`.